avocado.core.tree: Avoid cyclical dependency with avocado.core.tree_debug.
Merge the code of module avocado.core.tree_debug back to
to module avocado.core.tree, so we remove a cyclical dependency
between the modules.
